### 2023-11-07 #6026 @editorialbot
[PRE REVIEW]: autoStreamTree: Genomic variant data fitted to geospatial networks
**Submitting author:** @tkchafin (<a href="http://orcid.org/0000-0001-8687-5905">Tyler Chafin</a>)
**Repository:** https://github.com/tkchafin/autostreamtree
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
