### 2023-11-07 #6026 @Kevin-Mattheus-Moerman
@tkchafin thanks for this submission to JOSS. I am the AEiC for this track and here to help with initial steps. Below are some points that require your attention before we proceed:

- [ ] The above paper error seems to relate to the 5 rather than 4 zeros at the start of that ORCID. Fixing that should resolve this. Once you've fixed it you can call `@editorialbot generate pdf` to attempt to update/create the paper draft.
- [ ] Please check the above potentially invalid DOI :point_up:. You can try to address it and call `@editorialbot check references` here to check the DOIs again.
- [ ] Please extend the basic description of the project in your README, it currently only says `Implementation of Kalinowsky et al. 2008 'StreamTree' software for SNP datasets`. A short e.g. 1 or 2 paragraph description of the project would be great.
