### 2023-11-07 4beb4ec
remove invalid doi
