+++ paper/paper.bib
@@ -103,8 +103,7 @@
   journal={The Journal of Machine Learning Research},
   volume={12},
   pages={2825--2830},
-  year={2011},
-  doi={10.5555/1953048.2078195}
+  year={2011}
 }
 
 @article{Peterson2013,
