### 2023-12-04 #6026 @crvernon
:wave: @tkchafin I'll be your editor for this submission.  Once we get at least two reviewers on-board, I'll start the formal review.  In the meantime, I did notice that you do not have a test suite in your repository.  This is something that will come up in the main review.  Check out some other recently accepted JOSS publications that submitted Python packages if you are not familiar with building and automating tests.  Thanks!
