### 2023-12-06 7e27a50
updated installation to pip/ conda
