+++ setup.py
@@ -1,0 +1,40 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='autostreamtree',
+    version='1.0.1',
+    packages=find_packages(),
+    scripts=[
+        'scripts/networkDimensions.py',
+        'scripts/streeToDendrogram.py'
+    ],
+    entry_points={
+        'console_scripts': [
+            'autostreamtree=autostreamtree.cli:main',
+        ],
+    },
+    install_requires=[
+        "numpy",
+        "pandas>=2.0",
+        "networkx>=3.0",
+        "seaborn",
+        "matplotlib",
+        "geopandas",
+        "pyogrio",
+        "momepy",
+        "scipy",
+        "scikit-learn",
+        "mantel",
+        "pysam",
+        "sortedcontainers"
+    ],
+    extras_require={
+        'dev': ['pytest']
+    },
+    include_package_data=True,
+    author='Tyler Chafin',
+    author_email='tyler.chafin@bioss.ac.uk',
+    description='A package for fitting genetic distances to spatial networks',
+    keywords='population genetics, genetic distance, river networks',
+    url='https://github.com/tkchafin/autostreamtree',
+)
