+++ README.md
@@ -12,8 +12,11 @@ An input geospatial network (as shapefile, geodatabase, or .gpkg) is first used 
     1. [Interface](#ast_cli)
     2. [Inputs](#ast_inputs)
     3. [Genetic distances](#gendist)
-3. [Example analysis](#ast_example)
-4. [Outputs](#ast_outputs)
+    4. [Outputs](#ast_outputs)
+3. [Example](#tutorial)
+4. [Downstream analysis](#downstream)
+    1. [Utility scripts](#utils)
+    2. [Incorporating into workflows](#workflow)
 5. [References](#ast_refs)
 6. [Contributing Guidelines](#contrib)
 
@@ -22,7 +25,7 @@ An input geospatial network (as shapefile, geodatabase, or .gpkg) is first used 
 
 ### Dependencies
 
-autoStreamTree is a Python package which relies upon a number of packages:
+`autoStreamTree` is a Python package which relies upon a number of dependencies:
 
 ```
 - pandas
@@ -40,7 +43,7 @@ autoStreamTree is a Python package which relies upon a number of packages:
 - sortedcontainers
 ```
 
-autoStreamTree also uses methodology described in [Kalinowski et al. 2008](https://www.montana.edu/kalinowski/documents/2008_Stream_Trees%20_CJFAS.pdf), if you use this software you should cite: 
+`autoStreamTree` also uses methodology described in [Kalinowski et al. 2008](https://www.montana.edu/kalinowski/documents/2008_Stream_Trees%20_CJFAS.pdf), if you use this software you should cite: 
 ```
 Kalinowski ST, MH Meeuwig, SR Narum, ML Taper (2008) Stream trees: a statistical method for mapping genetic differences between populations of freshwater organisms to the sections of streams that connect them. Canadian Journal of Fisheries and Aquatic Sciences (65:2752-2760)
 ```
@@ -261,10 +264,10 @@ RAD_0   76      loc0_pos75      C       T       13      PASS    .       GT:DP:CA
 Parsing is faster with compressed and indexed files (and will be more space efficient), so I recommend keeping these compressed with bgzip and indexed with tabix:
 ```
 # compress
-bgzip test.vcf 
+bgzip test_sub100.vcf 
 
 # index
-tabix test.vcf.gz
+tabix test_sub100.vcf.gz
 ```
 
 This file should be provided using `-v/--vcf`.
@@ -309,7 +312,7 @@ or using an input file providing population assignments for each individual, aga
 #### Geodatabase 
 The input stream network can be provided as a shapefile, geodatabase, or GPKG file, all passed uing the `-s/--shp` option. There are a number of requirements for this file in order for the result to create a valid network. I highly recommend using the existing global stream datasets provided by the [HydroLab group](https://wp.geog.mcgill.ca/hydrolab/) at McGill University, specifically the [HydroAtlas](https://www.hydrosheds.org/page/hydroatlas) or [free-flowing rivers dataset](https://wp.geog.mcgill.ca/hydrolab/free-flowing-rivers/) as these are already properly formatted for use, and the additional variables included will make downstream analysis very easy. 
 
[…]
