### 2023-12-07 b439283
added endpoints for util scripts, updated README
