### 2023-12-07 3db38d3
styling and unit tests for sequence.py
