+++ tests/test_sequence.py
@@ -1,0 +1,162 @@
+import pytest
+import autostreamtree.sequence as seq_tools
+
+
+# Fixture for common reference and alternate alleles
+@pytest.fixture
+def ref_and_alts():
+    return "A", ["G", "C", "T"]
+
+
+# Parametrized test for basic decoding
+@pytest.mark.parametrize("gt, expected", [
+    ((1, 2), "G/C"),
+    ((0, 1), "A/G"),
+    ((2, 2), "C/C")
+])
+def test_basic_decode(gt, expected, ref_and_alts):
+    ref, alts = ref_and_alts
+    assert seq_tools.decode(gt, ref, alts) == expected
+
+
+# Test for handling missing genotypes
+def test_missing_genotype(ref_and_alts):
+    ref, alts = ref_and_alts
+    assert seq_tools.decode((None, None), ref, alts) == "N/N"
+
+
+# Parametrized test for IUPAC encoding
+@pytest.mark.parametrize("gt, expected", [
+    ((1, 2), seq_tools.dna_consensus("G/C")),
+    ((0, 1), seq_tools.dna_consensus("A/G")),
+    ((1, 1), seq_tools.dna_consensus("G/G")),
+])
+def test_iupac_encoding(gt, expected, ref_and_alts):
+    ref, alts = ref_and_alts
+    assert seq_tools.decode(gt, ref, alts, as_iupac=True) == expected
+
+
+# Test for tuple output
+def test_tuple_output(ref_and_alts):
+    ref, alts = ref_and_alts
+    gt = (1, 2)
+    assert seq_tools.decode(gt, ref, alts, as_tuple=True) == ("G", "C")
+
+
+# Test for list output
+def test_list_output(ref_and_alts):
+    ref, alts = ref_and_alts
+    gt = (1, 2)
+    assert seq_tools.decode(gt, ref, alts, as_list=True) == ["G", "C"]
+
+
+# Parametrized test for get_nuc_freqs function
+@pytest.mark.parametrize("seqs, ploidy, expected, raises_exception", [
+    # Scenario 1: Basic Frequency Calculation with No Ambiguities
+    (
+        {"sample1": "ACGT", "sample2": "TGCA", "sample3": "GTAC"},
+        1,
+        [{'a': 0.3333333333333333, 'g': 0.3333333333333333, 'c': 0.0,
+          't': 0.3333333333333333},
+         {'a': 0.0, 'g': 0.3333333333333333, 'c': 0.3333333333333333,
+          't': 0.3333333333333333},
+         {'a': 0.3333333333333333, 'g': 0.3333333333333333,
+          'c': 0.3333333333333333, 't': 0.0},
+         {'a': 0.3333333333333333, 'g': 0.0, 'c': 0.3333333333333333,
+          't': 0.3333333333333333}],
+        False
+    ),
+    # Scenario 2: Frequency Calculation with Ambiguities (Ploidy 1)
+    (
+        {"sample1": "AGR", "sample2": "TCY", "sample3": "GTS"},
+        1,
+        [{'a': 0.3333333333333333, 'g': 0.3333333333333333, 'c': 0.0,
+          't': 0.3333333333333333},
+         {'a': 0.0, 'g': 0.3333333333333333, 'c': 0.3333333333333333,
+          't': 0.3333333333333333},
+         {'a': 0, 'g': 0, 'c': 0, 't': 0}],
+        False
+    ),
+    # Scenario 3: Frequency Calculation with Ambiguities (Ploidy 2)
+    (
+        {"sample1": "AGR", "sample2": "TCY", "sample3": "GTS"},
+        2,
[…]
