+++ autostreamtree/sequence.py
@@ -1,23 +1,27 @@
-import os
-import sys
-
-from typing import List, Tuple, Dict, Any, Union, Optional
+from typing import List
 
 
-def decode(gt: tuple, ref: str, alts: list, as_iupac: bool = False, as_tuple: bool = False, as_list: bool = False) -> str:
+def decode(gt: tuple, ref: str, alts: list, as_iupac: bool = False,
+           as_tuple: bool = False, as_list: bool = False) -> str:
     """
-    Decode a genotype from a VCF file and return the corresponding DNA sequence.
+    Decode a genotype from a VCF file and return the corresponding DNA sequence
 
     Args:
-        gt (tuple): The genotype as a tuple of integers representing the indices of the reference and alternate alleles.
+        gt (tuple): The genotype as a tuple of integers representing the
+                    indices of the reference and alternate alleles.
         ref (str): The reference allele.
         alts (list): A list of alternate alleles.
-        as_iupac (bool): Whether to return the consensus IUPAC symbol if the genotype is heterozygous. Default is False.
-        as_tuple (bool): Whether to return the result as a tuple of two strings. Default is False.
-        as_list (bool): Whether to return the result as a list of two strings. Default is False.
+        as_iupac (bool): Whether to return the consensus IUPAC symbol if the
+                         genotype is heterozygous. Default is False.
+        as_tuple (bool): Whether to return the result as a tuple of two strings
+                         Default is False.
+        as_list (bool): Whether to return the result as a list of two strings.
+                        Default is False.
 
     Returns:
-        ret (str): The decoded genotype as a string. By default, this is a string in the format "ref/alt", but the format can be customized using the optional arguments.
+        ret (str): The decoded genotype as a string. By default, this is a
+                   string in the format "ref/alt", but the format can be
+                   customized using the optional arguments.
     """
     # initialize return value
     ret = [None, None]
@@ -52,38 +56,61 @@ def get_nuc_freqs(seqs: dict, ploidy: int) -> list:
     Compute the nucleotide frequencies of a set of DNA sequences.
 
     Args:
-        seqs (dict): A dictionary of DNA sequences. The keys are sample names and the values are strings of nucleotides.
-        ploidy (int): The ploidy of the sequences. If ploidy=1, ambiguities will be skipped. If ploidy=2, ambiguities will be resolved.
+        seqs (dict): A dictionary of DNA sequences. The keys are sample names
+                     and the values are strings of nucleotides.
+        ploidy (int): The ploidy of the sequences. If ploidy=1, ambiguities
+                      will be skipped. If ploidy=2, ambiguities will be
+                      resolved.
 
     Returns:
-        freqs (list): A list of dictionaries, where each dictionary contains the nucleotide frequencies for a single position in the sequences.
+        freqs (list): A list of dictionaries, where each dictionary contains
+                      the nucleotide frequencies for a single position in the
+                      sequences.
+
+    Raises:
+        ValueError: If any sequence is empty or contains only invalid
+                    characters, or if sequences are not all the same length.
     """
+    if not all(seqs):
+        raise ValueError("One or more sequences are empty.")
+
+    sequence_length = len(seqs[list(seqs.keys())[0]])
+    if not all(len(seq) == sequence_length for seq in seqs.values()):
+        raise ValueError("Sequences are not all the same length.")
+
     freqs = []
-    for loc in range(len(seqs[list(seqs.keys())[0]])):
[…]
