+++ autostreamtree/transform.py
@@ -1,160 +1,0 @@
-import sys
-import itertools
-import numpy as np
-import pandas as pd
-
-from typing import List, Tuple, Dict, Any, Union, Optional
-
-"""
-Transformations from Peterman et al. (2018) ResistanceGA package
-Full Citation:
-Peterman, W. E. 2018. ResistanceGA: An R package for the optimization 
-of resistance surfaces using genetic algorithms. Methods in Ecology 
-and Evolution doi:10.1111/2041-210X.12984
-GitHub:
-https://github.com/wpeterman/ResistanceGA/tree/3be50a6fa5515fe23953fca39135c5b340415607
-"""
-
-def rescale_cols(df: pd.DataFrame, m: float, M: float) -> pd.DataFrame:
-    """
-    Rescale all columns in a Pandas DataFrame to a new range between m and M.
-
-    Args:
-        df (pd.DataFrame): The DataFrame to rescale.
-        m (float): The minimum value of the new range.
-        M (float): The maximum value of the new range.
-
-    Returns:
-        rescaled_df (pd.DataFrame): A new DataFrame with all columns rescaled to the new range.
-    """
-    df -= df.min()
-    df /= df.max()
-    rescaled_df = (df * (M - m)) + m
-    return rescaled_df
-
-def ricker(dat: pd.Series, shape: float, ceiling: float) -> pd.Series:
-    """
-    Compute Ricker function values for a given range of data points.
-
-    Args:
-        dat (pd.Series): A pandas Series object containing the data points to compute Ricker values for.
-        shape (float): The "shape" parameter of the Ricker function.
-        ceiling (float): The "ceiling" parameter of the Ricker function.
-
-    Returns:
-        r (pd.Series): A pandas Series object containing the Ricker values for the input data points.
-    """
-    r = ceiling * dat * np.exp(-1 * dat / shape) + 1
-    return r
-
-def inv_ricker(dat: pd.Series, shape: float, ceiling: float) -> pd.Series:
-    """
-    Compute the inverse of the Ricker function for a given range of data points.
-
-    Args:
-        dat (pd.Series): A pandas Series object containing the data points to compute inverse Ricker values for.
-        shape (float): The "shape" parameter of the Ricker function.
-        ceiling (float): The "ceiling" parameter of the Ricker function.
-
-    Returns:
-        inv_r (pd.Series): A pandas Series object containing the inverse Ricker values for the input data points.
-    """
-    inv_r = (-1 * ceiling) * dat * np.exp(-1 * dat / shape) - 1
-    return inv_r
-
-def rev_inv_ricker(dat: pd.Series, shape: float, ceiling: float) -> pd.Series:
-    """
-    Compute the reverse inverse of the Ricker function for a given range of data points.
-
-    Args:
-        dat (pd.Series): A pandas Series object containing the data points to compute reverse inverse Ricker values for.
-        shape (float): The "shape" parameter of the Ricker function.
-        ceiling (float): The "ceiling" parameter of the Ricker function.
-
-    Returns:
-        rescaled_inv_r (pd.Series): A pandas Series object containing the reverse inverse Ricker values for the input data points, rescaled to a new range between the minimum and maximum values of the original values.
-    """
-    inv_r = inv_ricker(dat, shape, ceiling)
-    rescaled_inv_r = rescale_cols((-1 * inv_r), min(inv_r), max(inv_r))
-    return rescaled_inv_r
-
-def inv_ricker(dat: pd.Series, shape: float, ceiling: float) -> pd.Series:
-    """
-    Compute the reverse of the Ricker function for a given range of data points.
-
-    Args:
-        dat (pd.Series): A pandas Series object containing the data points to compute reverse Ricker values for.
[…]
