### 2023-12-07 b091887
fix deprecation warning, pandas>2
