### 2023-12-07 d1cb518
linting and tests for cluster_pops
