+++ autostreamtree/cluster_pops.py
@@ -1,53 +1,60 @@
-import os
-import sys
 import pandas as pd
 import numpy as np
 import seaborn as sns
+import matplotlib.pyplot as plt
 from sklearn.cluster import DBSCAN
 from shapely.geometry import MultiPoint
 from sortedcontainers import SortedDict
 
-from typing import List, Tuple, Dict, Any, Union, Optional
+from typing import Tuple, Dict, Optional
 
-import matplotlib.pyplot as plt
 
-#credit to Geoff Boeing at https://geoffboeing.com/2014/08/clustering-to-reduce-spatial-data-set-size/ for the tutorial
-
-#here coords should be a dictionary with key=ID and value=tuple of lat,long
-#epsilon=max distance (in km) between points in a cluster
-def dbscan_cluster(coords: Dict[str, Tuple[float, float]], epsilon: float, 
+# credit to Geoff Boeing at
+# https://geoffboeing.com/2014/08/clustering-to-reduce-spatial-data-set-size/
+# for the tutorial
+# here coords should be a dictionary with key=ID and value=tuple of lat,long
+# epsilon=max distance (in km) between points in a cluster
+def dbscan_cluster(coords: Dict[str, Tuple[float, float]], epsilon: float,
                    min_samples: int, out: Optional[str] = None) -> SortedDict:
     """
     Cluster geographic coordinates using DBSCAN algorithm.
 
     Parameters:
-    coords (Dict[str, Tuple[float, float]]): A dictionary of coordinate pairs with 
-                                             the key as the location identifier and 
-                                             values as tuples of (latitude, longitude).
-    epsilon (float): Maximum distance between two samples for them to be considered 
-                     as in the same neighborhood.
-    min_samples (int): The number of samples (or total weight) in a neighborhood 
-                       for a point to be considered as a core point.
+    coords (Dict[str, Tuple[float, float]]): A dictionary of coordinate pairs
+                                             with the key as the location
+                                             identifier and values as tuples of
+                                             (latitude, longitude).
+    epsilon (float): Maximum distance between two samples for them to be
+                     considered as in the same neighborhood.
+    min_samples (int): The number of samples (or total weight) in a
+                       neighborhood for a point to be considered as a core
+                       point.
     out (Optional[str]): Output file path to save cluster labels.
 
     Returns:
-    SortedDict: A sorted dictionary of cluster labels with keys as cluster name and 
-                values as lists of location identifiers in that cluster.
+    SortedDict: A sorted dictionary of cluster labels with keys as cluster name
+                and values as lists of location identifiers in that cluster.
 
     Raises:
     ValueError: If any coordinate value is missing or not a float.
 
     Examples:
-    >>> coords = {'A': (33.4484, -112.0740), 'B': (37.7749, -122.4194), 'C': (40.7128, -74.0060)}
+    >>> coords = {'A': (33.4484, -112.0740), 'B': (37.7749,
+                  -122.4194), 'C': (40.7128, -74.0060)}
     >>> dbscan_cluster(coords, 300, 2)
     SortedDict({'DB_-1': ['C'], 'DB_0': ['A'], 'DB_1': ['B']})
     """
     # Check if coordinates have correct values
     for k, v in coords.items():
         if not isinstance(v, tuple) or len(v) != 2:
[…]
