+++ tests/test_cluster_pops.py
@@ -1,0 +1,196 @@
+import pytest
+from unittest.mock import mock_open, patch
+import pandas as pd
+import numpy as np
+import matplotlib.pyplot as plt
+import autostreamtree.cluster_pops as cluster_pops
+from sortedcontainers import SortedDict
+
+
+# Test for basic functionality
+def test_dbscan_cluster_basic():
+    coords = {'A': (33.4484, -112.0740), 'B': (37.7749, -122.4194),
+              'C': (40.7128, -74.0060)}
+    epsilon = 300
+    min_samples = 2
+    expected_result = SortedDict({'DB_-1': ['A', 'B', 'C']})
+    assert cluster_pops.dbscan_cluster(coords, epsilon, min_samples) == \
+        expected_result
+
+
+# Test for invalid coordinates
+@pytest.mark.parametrize("coords, epsilon, min_samples", [
+    ({"A": (33.4484, None), "B": (37.7749, -122.4194)}, 300, 2),
+    ({"A": "invalid", "B": (37.7749, -122.4194)}, 300, 2),
+    # Add more cases as needed
+])
+def test_dbscan_cluster_invalid_coords(coords, epsilon, min_samples):
+    with pytest.raises(ValueError):
+        cluster_pops.dbscan_cluster(coords, epsilon, min_samples)
+
+
+# Test for output file writing
+@patch("builtins.open", new_callable=mock_open)
+def test_dbscan_cluster_output_file(mock_file):
+    coords = {'A': (33.4484, -112.0740), 'B': (37.7749, -122.4194)}
+    epsilon = 300
+    min_samples = 2
+    output_file = "test_output.txt"
+    cluster_pops.dbscan_cluster(coords, epsilon, min_samples, output_file)
+    mock_file.assert_called_with(output_file, 'w')
+
+
+def test_coords_to_dataframe():
+    coords = {'A': (33.4484, -112.0740), 'B': (37.7749, -122.4194)}
+    expected_df = pd.DataFrame(
+        {'long': [33.4484, 37.7749], 'lat': [-112.0740, -122.4194]},
+        index=['A', 'B']
+    )
+    pd.testing.assert_frame_equal(
+        cluster_pops.coords_to_dataframe(coords), expected_df)
+
+
+def test_coords_to_dataframe_empty():
+    coords = {}
+    # Create an empty DataFrame with an object index
+    expected_df = pd.DataFrame(
+        {'long': [], 'lat': []}, index=pd.Index([], dtype='object'))
+    expected_df = expected_df.astype({'long': 'float64', 'lat': 'float64'})
+    pd.testing.assert_frame_equal(
+        cluster_pops.coords_to_dataframe(coords), expected_df)
+
+
+def test_coords_to_matrix():
+    coords = {'A': (33.4484, -112.0740), 'B': (37.7749, -122.4194)}
+    expected_matrix = np.array([[33.4484, -112.0740], [37.7749, -122.4194]])
+    assert np.array_equal(
+        cluster_pops.coords_to_matrix(coords), expected_matrix)
+
+
+def test_coords_to_matrix_empty():
+    coords = {}
+    expected_matrix = np.array([], dtype=float).reshape(0, 2)
+    assert cluster_pops.coords_to_matrix(coords).shape == expected_matrix.shape
+
+
+def test_get_cluster_centroid_basic():
+    coords = SortedDict({'A': (33.4484, -112.0740), 'B': (37.7749, -122.4194)})
+    popmap = SortedDict({'cluster1': ['A'], 'cluster2': ['B']})
+    expected_centroids = SortedDict(
+        {'cluster1': (33.4484, -112.0740), 'cluster2': (37.7749, -122.4194)})
[…]
