### 2023-12-07 f0f20db
linting and tests for params.py
