+++ autostreamtree/params.py
@@ -1,25 +1,30 @@
 import sys
-import os
 import getopt
 
-#Object to parse command-line arguments
+
+# Object to parse command-line arguments
 class parseArgs():
     def __init__(self):
-        #Define options
+        # Define options
         try:
-            options, remainder = getopt.getopt(sys.argv[1:], 'hs:i:r:p:d:a:lw:o:gP:L:Scn:G:v:C:O:', \
-            ["shp=", "help", "input=", "run=", "pop=", "popmap=","dist=", "agg_method=",
-            "het", "genmat=", "snp", "snps", "msat", "msats", "log", "and_log", "iterative",
-            "weight=", "out=", "method=", "plots", "plot","perm=", "phased", "median",
-            "diploid", "geopop", "geopops", "global_het", "haploid", "loc_agg=",
-            "pop_agg=", "sdist_agg=", "clusterpop", "epsilon=", "min_samples=", "sclusterpop",
-            "network=", "overwrite", "reachid_col=", "length_col=", "coercemat", "locmatdir=",
-            "vcf=", "concat=", "edge_list=", "gdf_out="])
+            options, _ = getopt.getopt(
+                sys.argv[1:], 'hs:i:r:p:d:a:lw:o:gP:L:Scn:G:v:C:O:',
+                ["shp=", "help", "input=", "run=", "pop=", "popmap=", "dist=",
+                 "agg_method=", "het", "genmat=", "snp", "snps", "msat",
+                 "msats", "log", "and_log", "iterative", "weight=", "out=",
+                 "method=", "plots", "plot", "perm=", "phased", "median",
+                 "diploid", "geopop", "geopops", "global_het", "haploid",
+                 "loc_agg=", "pop_agg=", "sdist_agg=", "clusterpop",
+                 "epsilon=", "min_samples=", "sclusterpop", "network=",
+                 "overwrite", "reachid_col=", "length_col=", "coercemat",
+                 "locmatdir=", "vcf=", "concat=", "edge_list=", "gdf_out="]
+            )
         except getopt.GetoptError as err:
             print(err)
-            self.display_help("\nExiting because getopt returned non-zero exit status.")
-        #Default values for params
-        #Input params
+            self.display_help(
+                "\nExiting because getopt returned non-zero exit status.")
+        # Default values for params
+        # Input params
         self.shapefile = None
         self.locmatdir = None
         self.geodb = None
@@ -30,9 +35,9 @@ class parseArgs():
         self.pop = False
         self.overwrite = False
         self.geopop = False
-        self.clusterpop=False
-        self.sclusterpop=False
-        self.output_driver="GPKG"
+        self.clusterpop = False
+        self.sclusterpop = False
+        self.output_driver = "GPKG"
         self.dist = "LINFST"
         self.het = False
         self.genmat = None
@@ -43,59 +48,67 @@ class parseArgs():
         self.permutations = 1000
         self.coercemat = False
         self.method = "PEARSON"
-        self.edge_list=None
-        self.plots=False
-        self.out="out"
-        self.median=False
-        self.ploidy=2
-        self.global_het=False
+        self.edge_list = None
+        self.plots = False
+        self.out = "out"
+        self.median = False
+        self.ploidy = 2
+        self.global_het = False
         self.loc_agg = "ARITH"
         self.pop_agg = "ARITH"
-        self.sdist_agg="ARITH"
+        self.sdist_agg = "ARITH"
 
         self.reachid_col = "HYRIV_ID"
         self.length_col = "LENGTH_KM"
 
-        #dbscan Options
-        self.min_samples=1
-        self.epsilon=20
+        # dbscan Options
+        self.min_samples = 1
+        self.epsilon = 20
 
-
[…]

[…]
