+++ autostreamtree/genetic_distances.py
@@ -1,169 +1,194 @@
 import sys
-import os
 import itertools
-import math
 import scipy
 import numpy as np
-import pandas as pd
-from sortedcontainers import SortedDict
-
-from typing import List, Tuple, Dict, Any, Union, Optional
+from typing import List
 
 import autostreamtree.aggregators as agg
 import autostreamtree.sequence as seq
 
-def get_pop_genmat(dist, indmat, popmap, dat, seqs, pop_agg="ARITH", loc_agg="ARITH", ploidy=2, global_het=False):
-    #make matrix
-    genmat = np.zeros((len(popmap),len(popmap)))
-    #establish as nan
+
+def get_pop_genmat(dist, indmat, popmap, dat, seqs, pop_agg="ARITH",
+                   loc_agg="ARITH", ploidy=2, global_het=False):
+    # make matrix
+    genmat = np.zeros((len(popmap), len(popmap)))
+    # establish as nan
     genmat[:] = np.nan
-    #for each combination, either average ind distances or calc freq dist
-    for ia,ib in itertools.combinations(range(0,len(popmap)),2):
-        if dist in ["JC69", "K2P", "PDIST", "TN84", "TN93"]:
-            #print(popmap.keys()[ia], popmap.values()[ia])
+    # for each combination, either average ind distances or calc freq dist
+    for ia, ib in itertools.combinations(range(0, len(popmap)), 2):
+        if dist in ["JC69", "PDIST"]:
             inds1 = [dat.index(x) for x in popmap.values()[ia]]
             inds2 = [dat.index(x) for x in popmap.values()[ib]]
-            #print(inds1)
-            genmat[ia,ib] = (agg.aggregate_dist(pop_agg, ([indmat[i, j] for i in inds1 for j in inds2])))
-            genmat[ib,ia] = genmat[ia,ib]
-        elif dist == "JOST":
-            results=list()
-            for loc in range(0, len(seqs[popmap.values()[ia][-1]])):
-                seqs1 = get_alleles([seqs[x][loc] for x in popmap.values()[ia]])
-                seqs2 = get_alleles([seqs[x][loc] for x in popmap.values()[ib]])
-                if not clean_list(set(seqs1), ["n", "N", "-", "?"]) or not clean_list(set(seqs2), ["n", "N", "-", "?"]):
-                    continue
-                results.append(two_pop_jost_d(seqs1, seqs2, ploidy, global_het))
-            if len(results) > 1:
-                genmat[ia,ib] = genmat[ib,ia] = agg.aggregate_dist(loc_agg, results)
-            elif len(results) < 1:
-                #print("ERROR: population",popmap.values()[ia],"or",popmap.values()[ib],"lacks any data")
-                raise ValueError
-            else:
-                genmat[ia,ib] = genmat[ib,ia] = results[0]
-        elif dist == "GST" or dist == "GSTPRIME":
-            HT=list()
-            HS=list()
-            for loc in range(0, len(seqs[popmap.values()[ia][-1]])):
-                seqs1 = get_alleles([seqs[x][loc] for x in popmap.values()[ia]])
-                seqs2 = get_alleles([seqs[x][loc] for x in popmap.values()[ib]])
-                if not clean_list(set(seqs1), ["n", "N", "-", "?"]) or not clean_list(set(seqs2), ["n", "N", "-", "?"]):
-                    continue
-                if dist == "GST" or "GSTPRIME":
-                    (ht, hs) = two_pop_ht_hs(seqs1, seqs2, ploidy, global_het)
-                    HT.append(ht)
-                    HS.append(hs)
-            Ht_global = np.mean(HT)
-            Hs_global = np.mean(HS)
-            if dist == "GST":
-                if Ht_global <= 0.0:
-                    genmat[ia,ib] = genmat[ib,ia] = 0.0
[…]
