+++ autostreamtree/genetic_distances.py
@@ -1,11 +1,21 @@
 import sys
 import itertools
-import scipy
 import numpy as np
 from typing import List
 
 import autostreamtree.aggregators as agg
 import autostreamtree.sequence as seq
+
+
+"""
+Note several distance calculations not currently in use, I plan to expand this
+further in the future, but for now only the following are supported:
+- p distances
+- JC69
+- Weir and Cockerham's Fst (linearised or non-linearised)
+- Jost's D
+- Chord distance
+"""
 
 
 def get_pop_genmat(dist, indmat, popmap, dat, seqs, pop_agg="ARITH",
@@ -29,8 +39,10 @@ def get_pop_genmat(dist, indmat, popmap, dat, seqs, pop_agg="ARITH",
             gn = len(popmap)
 
             for loc in range(0, len(seqs[popmap.values()[ia][-1]])):
-                seqs1 = clean_inds([seqs[x][loc] for x in popmap.values()[ia]])
-                seqs2 = clean_inds([seqs[x][loc] for x in popmap.values()[ib]])
+                seqs1 = get_alleles(
+                    [seqs[x][loc] for x in popmap.values()[ia]])
+                seqs2 = get_alleles(
+                    [seqs[x][loc] for x in popmap.values()[ib]])
                 if (not clean_list(set(seqs1), ["n", "N", "-", "?"]) or
                         not clean_list(set(seqs2), ["n", "N", "-", "?"])):
                     continue
@@ -45,6 +57,7 @@ def get_pop_genmat(dist, indmat, popmap, dat, seqs, pop_agg="ARITH",
             # Aggregate the results for Hs and Ht
             global_hs = agg.aggregate_dist(loc_agg, hs_vals)
             global_ht = agg.aggregate_dist(loc_agg, ht_vals)
+            hs_vals.append(global_hs)
 
             # Calculate global D
             if dist == "JOST":
@@ -165,7 +178,7 @@ def get_pop_genmat(dist, indmat, popmap, dat, seqs, pop_agg="ARITH",
     return genmat
 
 
-# function computes pairwise JC69-corrected genetic distances
+# function computes pairwise p-distance or JC69-corrected distances
 def get_genmat(dist, points, seqs, ploidy, het, loc_agg):
     # make matrix
     genmat = np.zeros((len(points), len(points)))
@@ -365,7 +378,6 @@ def jukes_cantor_distance(seq1: str, seq2: str, het: bool = False) -> float:
 #     return dist
 
 
-
 # p distance = D / L (differences / length)
 # L is the UNGAPPED distance
 # ambigs are expanded
@@ -373,28 +385,28 @@ def jukes_cantor_distance(seq1: str, seq2: str, het: bool = False) -> float:
 # P (transitions/L) and Q (transversions/L)
 def p_distance(seq1, seq2, trans=False, transSplit=False):
     L = min(len(seq1), len(seq2))
-    D=0.0
-    P=0.0
-    Q=0.0
-    P1=0.0
-    P2=0.0
+    D = 0.0
+    P = 0.0
+    Q = 0.0
+    P1 = 0.0
+    P2 = 0.0
     for n1, n2 in zip(seq1.lower(), seq2.lower()):
         if n1 in ["?", "-", "n"] or n2 in ["?", "-", "n"] :
             L = L-1
             continue
         elif n1 == n2:
             continue
-        else: #if n1 and n2 not equal and not gaps
+        else:  # if n1 and n2 not equal and not gaps
             if n1 in ["a", "c", "g", "t"] and n2 in ["a", "c", "g", "t"]:
[…]
