### 2023-12-07 bea0a60
simplified p-distance calculations
