+++ tests/test_genetic_distances.py
@@ -1,0 +1,190 @@
+import numpy as np
+import pytest
+from sortedcontainers import SortedDict
+from unittest.mock import patch
+import autostreamtree.genetic_distances as gendist
+
+
+dist_options = ["JC69", "PDIST", "JOST", "HARMD", "FST", "LINFST", "CHORD"]
+
+
+def mock_agg_dist(*args, **kwargs):
+    # Mock behavior of agg.aggregate_dist
+    return 0.5  # Return a constant value or based on input for testing
+
+
+@pytest.fixture
+def mock_popmap():
+    # Example mock population map
+    return SortedDict(
+        {'Pop1': ['Ind1', 'Ind2'], 'Pop2': ['Ind3', 'Ind4']}
+    )
+
+
+@pytest.fixture
+def mock_dat():
+    # Example mock data
+    return ['Ind1', 'Ind2', 'Ind3', 'Ind4']
+
+
+@pytest.fixture
+def mock_points():
+    return SortedDict(
+        {'Ind1': (0, 0), 'Ind2': (1, 0), 'Ind3': (0, 1), 'Ind4': (1, 1)}
+    )
+
+
+@pytest.fixture
+def mock_snps():
+    return {'Ind1': ['A/A', 'A/T', 'A/A'], 'Ind2': ['A/T', 'A/T', 'A/A'],
+            'Ind3': ['A/T', 'A/T', 'T/T'], 'Ind4': ['T/T', 'A/T', 'T/T']}
+
+
+@pytest.fixture
+def mock_seqs():
+    return {
+        'Ind1': ['AAAAAAA', 'AAAAAAT', 'AAAAAAA'],
+        'Ind2': ['AAAAAAT', 'AAAAAAT', 'AAAAAAA'],
+        'Ind3': ['AAAAAAT', 'AAAAAAT', 'TTTTTTT'],
+        'Ind4': ['TTTTTTT', 'AAAAAAT', 'TTTTTTT']
+    }
+
+
+@pytest.fixture
+def mock_seqs_het():
+    return {
+        'Ind1': ['AAAAAAA/AAAAAAA', 'AAAAAAA/AAAAAAA'],
+        'Ind2': ['AAAAAAA/AAAAAAT', 'AAAAAAA/AAAAAAT'],
+        'Ind3': ['AAAAAAA/TTTTTTT', 'AAAAAAA/TTTTTTT'],
+        'Ind4': ['TTTTTTT/TTTTTTT', 'TTTTTTT/TTTTTTT']
+    }
+
+
+@pytest.fixture
+def mock_indmat():
+    # Example mock individual distance matrix
+    return np.array([[0, 1, 2, 3], [1, 0, 4, 5], [2, 4, 0, 6], [3, 5, 6, 0]])
+
+
+@patch('autostreamtree.genetic_distances.agg.aggregate_dist',
+       side_effect=mock_agg_dist)
+def test_get_pop_genmat_unsupported_option(mock_agg, mock_popmap, mock_dat,
+                                           mock_snps, mock_indmat):
+    with pytest.raises(ValueError):
+        gendist.get_pop_genmat("UNSUPPORTED_DIST", mock_indmat, mock_popmap,
+                               mock_dat, mock_snps)
+
+
+@pytest.mark.parametrize("dist", dist_options)
+@patch('autostreamtree.genetic_distances.agg.aggregate_dist',
+       side_effect=mock_agg_dist)
+def test_get_pop_genmat_valid_options(mock_agg, mock_popmap, mock_dat,
+                                      mock_snps, mock_indmat, dist):
+    result = gendist.get_pop_genmat(dist, mock_indmat, mock_popmap, mock_dat,
+                                    mock_snps)
+
+    assert isinstance(result, np.ndarray)
[…]
