### 2023-12-08 ffc533c
doc changes; more tests
