### 2023-12-08 ed7c993
styling and tests for genetic_distances
