+++ autostreamtree/genetic_distances.py
@@ -1,7 +1,6 @@
 import sys
 import itertools
 import numpy as np
-from typing import List
 
 import autostreamtree.aggregators as agg
 import autostreamtree.sequence as seq
@@ -25,7 +24,7 @@ def get_pop_genmat(dist, indmat, popmap, dat, seqs, pop_agg="ARITH",
     genmat[:] = np.nan
     # for each combination, either average ind distances or calc freq dist
     for ia, ib in itertools.combinations(range(0, len(popmap)), 2):
-        if dist in ["JC69", "PDIST"]:
+        if dist in ["PDIST"]:
             inds1 = [dat.index(x) for x in popmap.values()[ia]]
             inds2 = [dat.index(x) for x in popmap.values()[ib]]
             genmat[ia, ib] = (agg.aggregate_dist(
@@ -135,23 +134,23 @@ def get_pop_genmat(dist, indmat, popmap, dat, seqs, pop_agg="ARITH",
                     genmat[ia, ib] = genmat[ib, ia] = (theta / (1-theta))
                 else:
                     genmat[ia, ib] = genmat[ib, ia] = theta
-        elif dist == "NEI83":
-            results = list()
-            loci = 0.0
-            for loc in range(0, len(seqs[popmap.values()[ia][-1]])):
-                seqs1 = get_alleles(
-                    [seqs[x][loc] for x in popmap.values()[ia]]
-                )
-                seqs2 = get_alleles(
-                    [seqs[x][loc] for x in popmap.values()[ib]]
-                )
-                if (not clean_list(set(seqs1), ["n", "N", "-", "?"]) or
-                        not clean_list(set(seqs2), ["n", "N", "-", "?"])):
-                    continue
-                loci += 1.0
-                results.append(two_pop_nei_da(seqs1, seqs2))
-            Da = (1.0 - (np.sum(results) / loci))
-            genmat[ia, ib] = genmat[ib, ia] = Da
+        # elif dist == "NEI83":
+        #     results = list()
+        #     loci = 0.0
+        #     for loc in range(0, len(seqs[popmap.values()[ia][-1]])):
+        #         seqs1 = get_alleles(
+        #             [seqs[x][loc] for x in popmap.values()[ia]]
+        #         )
+        #         seqs2 = get_alleles(
+        #             [seqs[x][loc] for x in popmap.values()[ib]]
+        #         )
+        #         if (not clean_list(set(seqs1), ["n", "N", "-", "?"]) or
+        #                 not clean_list(set(seqs2), ["n", "N", "-", "?"])):
+        #             continue
+        #         loci += 1.0
+        #         results.append(two_pop_nei_da(seqs1, seqs2))
+        #     Da = (1.0 - (np.sum(results) / loci))
+        #     genmat[ia, ib] = genmat[ib, ia] = Da
         elif dist == "CHORD":
             sum_squares = 0.0
             for loc in range(0, len(seqs[popmap.values()[ia][-1]])):
@@ -208,13 +207,13 @@ def get_genmat(dist, points, seqs, ploidy, het, loc_agg):
                 seq1 = seq.dna_consensus(seq1)
             if "/" in seq2:
                 seq2 = seq.dna_consensus(seq2)
-            if dist == "JC69":
-                results.append(jukes_cantor_distance(seq1, seq2, het))
-            elif dist == "PDIST":
+            if dist == "PDIST":
                 if het:
                     results.append(p_distance(seq1, seq2))
                 else:
                     results.append(hamming_distance(seq1, seq2))
+            else:
+                raise ValueError(f"{dist} not implemented")
[…]
