### 2023-12-08 ef423c3
more tests, linting, and organisation
