+++ autostreamtree/functions.py
@@ -1,14 +1,10 @@
-
 import sys
 import os
-
-os.environ['USE_PYGEOS'] = '0'
-
 import itertools
 import traceback
 import math
-import getopt
 import momepy
+import warnings
 import pyogrio
 import pandas as pd
 import numpy as np
@@ -16,33 +12,30 @@ import networkx as nx
 import seaborn as sns
 from pysam import VariantFile
 from scipy import stats
-from os import listdir
-from os.path import isfile, join
 from sortedcontainers import SortedDict
-from sklearn.linear_model import LinearRegression
 from networkx import NodeNotFound
 import matplotlib.pyplot as plt
 import pickle
 import mantel
-from math import radians, degrees, sin, cos, asin, acos, sqrt
+from math import radians, sin, cos, acos
 
 import autostreamtree.cluster_pops as clust
 import autostreamtree.sequence as seq
 import autostreamtree.genetic_distances as gendist
-import autostreamtree.report_refs as ref
-import autostreamtree.aggregators as agg
 
-from typing import List, Tuple, Dict, Any, Union, Optional
 
 # suppress irrelevant warning from momepy
-import warnings
-def custom_warn_handler(message, category, filename, lineno, file=None, line=None):
+def custom_warn_handler(message, category, filename, lineno, file=None,
+                        line=None):
     if "momepy/utils.py" in filename and issubclass(category, UserWarning):
         return
-    return original_showwarning(message, category, filename, lineno, file, line)
+    return original_showwarning(message, category, filename, lineno, file,
+                                line)
+
 
 original_showwarning = warnings.showwarning
 warnings.showwarning = custom_warn_handler
+
 
 def read_vcf(vcf, concat="none", popmap=None):
     """
@@ -50,14 +43,17 @@ def read_vcf(vcf, concat="none", popmap=None):
 
     ARgs:
         vcf: Path to the input VCF file.
-        concat: Specifies the concatenation mode for genotypes. Options are "all", "loc", and "none".
+        concat: Specifies the concatenation mode for genotypes. Options are
+                "all", "loc", and "none".
                    "all": Concatenate genotypes of all loci for each sample.
-                   "loc": Concatenate genotypes within the same chromosome for each sample.
+                   "loc": Concatenate genotypes within the same chromosome for
+                          each sample.
                    "none": Do not concatenate genotypes.
-        popmap: Optional dictionary that maps populations to a list of samples. If provided,
-                   only samples in the popmap will be retained in the output dictionary.
+        popmap: Optional dictionary that maps populations to a list of samples.
+                If provided, only samples in the popmap will be retained in the
+                output dictionary.
     Returns:
-        A dictionary with sample names as keys and lists of genotypes as values.
+        A dictionary with sample names as keys and lists of genotypes as values
     """
 
     bcf_in = VariantFile(vcf)
@@ -66,12 +62,12 @@ def read_vcf(vcf, concat="none", popmap=None):
     vcf_samples = list(bcf_in.header.samples)
 
     # set up data dict
-    dat=dict()
+    dat = dict()
     samples = list((bcf_in.header.samples))
     for s in samples:
         if concat == "all":
             dat[s] = list()
-            dat[s].append(["",""])
+            dat[s].append(["", ""])
         else:
             dat[s] = list()
 
@@ -83,23 +79,26 @@ def read_vcf(vcf, concat="none", popmap=None):
         keep = [s for s in keep if s in vcf_samples]
         bcf_in.subset_samples(keep)
 
-    chrom="FIRST"
+    chrom = "FIRST"
     for record in bcf_in.fetch():
         for i, sample in enumerate(record.samples):
-            if concat=="all":
[…]
