+++ autostreamtree/cli.py
@@ -1,13 +1,11 @@
 import sys
-import os
-
-os.environ['USE_PYGEOS'] = '0'
-
 import pandas as pd
 import numpy as np
 import geopandas as gpd
 import pyogrio
+import random
 import matplotlib.pyplot as plt
+import time
 
 import autostreamtree.functions as ast
 from autostreamtree.params import parseArgs
@@ -15,17 +13,25 @@ import autostreamtree.cluster_pops as clust
 import autostreamtree.report_refs as ref
 import autostreamtree.aggregators as agg
 
+
 def main():
 
     params = parseArgs()
 
-    #########################################################
-    # Step 1: Reading in datasets
-    #########################################################
+    # Set the seed for random number generation
+    if params.seed is not None:
+        np.random.seed(params.seed)
+        random.seed(params.seed)
+    else:
+        # Use clock time to seed RNG if params.seed is None
+        clock_seed = int(time.time() * 1000)
+        np.random.seed(clock_seed)
+        random.seed(clock_seed)
 
+    # Step 1: Reading in datasets
     print("Running autostream tree", params.run, "workflow\n")
     G = ast.read_network(params.network, params.shapefile)
-    # prune graph if list of edges passed 
+    # prune graph if list of edges passed
     if params.edge_list:
         df = pd.read_csv(params.edge_list, sep='\t')
         edge_list = df[params.reachid_col].tolist()
@@ -39,22 +45,23 @@ def main():
     if params.run != "STREAMDIST":
         seqs = ast.read_vcf(params.vcf, concat=params.concat, popmap=popmap)
 
-    #########################################################
     # Step 2: Calculating genetic distance matrix(-ces)
     # Optional: Calculate separately for each locus
-    #########################################################
 
-    gen=None
-    pop_gen=None
+    gen = None
+    pop_gen = None
     # calculate genetic distance matrix
     if params.run != "STREAMDIST" and params.run != "RUNLOCI":
         if not params.genmat:
             print("\nCalculating genetic distances...")
-            (gen, pop_gen) = ast.get_gendist_mats(params, point_coords, popmap, seqs)
+            (gen, pop_gen) = ast.get_gendist_mats(params, point_coords, popmap,
+                                                  seqs)
         else:
-            print("\nReading genetic distances from provided matrix:", params.genmat)
+            print("\nReading genetic distances from provided matrix:",
+                  params.genmat)
             inmat = pd.read_csv(params.genmat, header=0, index_col=0, sep="\t")
-            (gen, pop_gen) = ast.parse_input_genmat(params, inmat, point_coords, popmap)
+            (gen, pop_gen) = ast.parse_input_genmat(params, inmat,
+                                                    point_coords, popmap, seqs)
 
         ast.report_genmats(params, gen, pop_gen, point_coords, pop_coords)
 
@@ -62,77 +69,83 @@ def main():
         if params.run == "GENDIST":
             sys.exit(0)
 
-    #IMPORTANT NOTE: This function will silently skip any loci for which calculations aren't possible (e.g., population having no data)
+    # IMPORTANT NOTE: This function will silently skip any loci for which
+    # calculations aren't possible (e.g., population having no data)
     if params.run == "RUNLOCI":
-        genlist=list()
-        popgenlist=list()
+        genlist = list()
+        popgenlist = list()
         for loc in ast.get_loc_data(seqs):
             try:
-                (gen, pop_gen) = ast.get_gendist_mats(params, point_coords, popmap, loc)
+                (gen, pop_gen) = ast.get_gendist_mats(params, point_coords,
+                                                      popmap, loc)
                 genlist.append(gen)
                 popgenlist.append(pop_gen)
[…]
