+++ tests/test_import.py
@@ -1,0 +1,12 @@
+import importlib
+import pkgutil
+
+
+def test_import_all_modules():
+    # Import all submodules in the resistnet package
+    package_name = 'autostreamtree'
+    package = importlib.import_module(package_name)
+
+    for importer, modname, ispkg in pkgutil.walk_packages(
+            package.__path__, package_name + '.'):
+        importlib.import_module(modname)
