### 2023-12-11 d7e0e3f
unit tests and linting
