+++ tests/test_functions.py
@@ -1,7 +1,80 @@
 import os
 import pytest
+from io import StringIO
+import sys
+import tempfile
+import numpy as np
+import pandas as pd
+import pickle
+import networkx as nx
+from scipy import stats
+from sortedcontainers import SortedDict
+from unittest.mock import patch, mock_open, MagicMock
+import geopandas as gpd
 import autostreamtree
 import autostreamtree.functions as funcs
+
+
+@pytest.fixture
+def sample_distances():
+    return np.array([[0, 1, 1],
+                     [1, 0, 2],
+                     [3, 1, 1]])
+
+
+@pytest.fixture
+def sample_incidence_matrix():
+    return np.array([[0, 0, 1],
+                     [0, 1, 0],
+                     [1, 0, 0]])
+
+
+@pytest.fixture
+def empty_graph():
+    return nx.Graph()
+
+
+@pytest.fixture
+def sample_graph():
+    G = nx.Graph()
+    G.add_edge(1, 2, weight=10, reachid_col='a')
+    G.add_edge(2, 3, weight=20, reachid_col='b')
+    G.add_edge(3, 4, weight=30, reachid_col='c')
+    return G
+
+
+@pytest.fixture
+def sample_gen_matrix():
+    return np.array([[0, 2, 3], [2, 0, 1], [3, 1, 0]])
+
+
+@pytest.fixture
+def sample_pop_gen_matrix():
+    return np.array([[0, 1], [1, 0]])
+
+@pytest.fixture
+def sample_geo_matrix():
+    return np.array([[0, 1, 2], [1, 0, 3], [2, 3, 0]])
+
+
+@pytest.fixture
+def mock_network_path():
+    G = nx.Graph()
+    G.add_edge('A', 'B')
+    G.add_edge('B', 'C')
+    with tempfile.NamedTemporaryFile(mode='wb', suffix='.pickle', delete=False) as temp_file:
+        pickle.dump(G, temp_file)
+        return temp_file.name
+
+
+@pytest.fixture
+def point_coords():
+    return {'A': (0, 0), 'B': (1, 1), 'C': (2, 2)}
+
+
+@pytest.fixture
+def pop_coords():
+    return {'Pop1': (0, 0), 'Pop2': (1, 1)}
 
 
 @pytest.fixture
@@ -12,7 +85,454 @@ def shapefile_path():
 
 
 @pytest.fixture
-def coords_path():
-    base_path = os.path.dirname(autostreamtree.__file__)
-    file_path = os.path.join(base_path, 'data', 'test.pointCoords.txt')
-    return file_path
+def vcf_file_path():
+    with tempfile.NamedTemporaryFile(
+            mode='w', suffix='.vcf', delete=False) as temp_file:
+        temp_file.write("""##fileformat=VCFv4.2
+#CHROM\tPOS\tID\tREF\tALT\tQUAL\tFILTER\tINFO\tFORMAT\tInd1\tInd2\tInd3\tInd4
+1\t100\t.\tA\tT\t.\t.\t.\tGT\t0/0\t0/1\t0/1\t1/1
+1\t150\t.\tA\tT\t.\t.\t.\tGT\t0/1\t0/1\t0/1\t0/1
+2\t200\t.\tA\tT\t.\t.\t.\tGT\t0/0\t0/0\t1/1\t1/1
+""")
+        return temp_file.name
+
+
+@pytest.fixture
+def mock_points_df():
+    data = {'sample': ['Ind1', 'Ind2', 'Ind3', 'Ind4'],
+            'lat': [26.925414, 26.925414, 27.21041, 27.21041],
+            'long': [90.396139, 90.396139, 91.20795, 91.20795]}
[…]
