+++ conda-recipe/meta.yaml
@@ -6,7 +6,7 @@ channels:
 
 package:
   name: autostreamtree
-  version: "1.0.1"
+  version: "1.1.0"
 
 source:
   path: ../ 
