### 2023-12-11 e86d8af
set up github CI to run tests
