### 2023-12-11 #6026 @tkchafin
@crvernon Thanks for the heads up. I've added an automated test suite to the repository, set to run on push/pull!
