### 2023-12-18 #6160 @editorialbot
[REVIEW]: autoStreamTree: Genomic variant data fitted to geospatial networks
**Submitting author:** @tkchafin (<a href="http://orcid.org/0000-0001-8687-5905">Tyler Chafin</a>)
**Repository:** https://github.com/tkchafin/autostreamtree
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.5
