### 2024-01-09 tkchafin/autostreamtree#5 @xin-huang
Furthermore, considering that Apple Silicon (https://github.com/actions/runner-images/issues/8439) is available in GitHub Actions now, it would be beneficial for the authors to include builds for OSX arm64 in the [ci.yml](https://github.com/tkchafin/autostreamtree/blob/main/.github/workflows/ci.yml). This addition would aid in verifying the effectiveness of the installation instructions for OSX arm64.
