### 2024-01-10 tkchafin/autostreamtree#6 @xin-huang
[JOSS REVIEW] Suggestions for Software Paper
This is a part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/6160)

1. **A statement of need**

In the provided text, it seems the phrase "reach- level within the spatial network" should be corrected to "river- level within the spatial network."

https://github.com/tkchafin/autostreamtree/blob/e86d8af440cf1c614b64f717cf9078d49f06fba5/paper/paper.md?plain=1#L42

2. **State of the field**

It is noted that a "State of the field" section is absent in the manuscript. As this is a requirement for the journal, it is recommended that the authors include this section to provide context and relevance of the work in the current research landscape.

3. **Quality of writing**

In the following text, the hyphens in the words "de-rived" and "correla-tions" should be removed for proper spelling.

https://github.com/tkchafin/autostreamtree/blob/e86d8af440cf1c614b64f717cf9078d49f06fba5/paper/paper.md?plain=1#L50

A tilde is missing in the term `(linearized *F*~ST)` from the following text.

https://github.com/tkchafin/autostreamtree/blob/e86d8af440cf1c614b64f717cf9078d49f06fba5/paper/paper.md?plain=1#L72

The correct notation should be `(linearized *F*~ST~)`

Regarding the mention of the U.S. Fish & Wildlife Service in the acknowledgment section, it would be clearer if the authors either provide direct links to the relevant websites or explicitly state that they are acknowledging the U.S. Fish & Wildlife Service without providing links.

For Figure 1, I recommend adding a label "F_ST distances" adjacent to the color bar to enhance reader comprehension of the color coding in Figure 1A.

4. **References**

The reference citations `[Hagberg2008]` and `[Weir1984]` are missing the "@" symbol. They should be corrected to `[@Hagberg2008]` and `[@Weir1984]` in the following text, respectively.

https://github.com/tkchafin/autostreamtree/blob/e86d8af440cf1c614b64f717cf9078d49f06fba5/paper/paper.md?plain=1#L48

https://github.com/tkchafin/autostreamtree/blob/e86d8af440cf1c614b64f717cf9078d49f06fba5/paper/paper.md?plain=1#L58

The reference `Hagberg2008` is not listed in the reference section. It should be added to maintain citation accuracy. For the VCF format, the authors should consider citing the paper by [Danecek et al. 2011](https://doi.org/10.1093/bioinformatics/btr330) to provide a comprehensive background and reference for this format.
