### 2024-01-10 #6160 @xin-huang
The authors have developed a valuable tool for the study of landscape genetics utilizing population genomic data. Given the increasing availability of genomic and geospatial data, this tool stands to aid researchers in understanding how evolutionary or ecological processes influence the spatial patterns of genetic variation. I have raised specific concerns regarding the documentation and the software paper through the opening of two issues https://github.com/tkchafin/autostreamtree/issues/5 and https://github.com/tkchafin/autostreamtree/issues/6. Additionally, it is imperative that the authors make the data and codes available for replicating the analyses and performance comparisons presented in their paper, as this is a requirement of the journal.
