### 2024-01-18 tkchafin/autostreamtree#5 @radeva
> it would be beneficial for the authors to include builds for OSX arm64 in the [ci.yml](https://github.com/tkchafin/autostreamtree/blob/main/.github/workflows/ci.yml).

@tkchafin Feel free to use [FlyCI](https://www.flyci.net/)'s M1 and M2 runners. Our runners are on average 2x faster and 2x cheaper than GitHub's **AND** we have a free tier for OSS projects (see below).

# Install Instructions

1. Install [FlyCI app](https://github.com/apps/flyci-prod) and
2. Use a FlyCI runner in your workflow:

```diff
jobs:
 ci:
-    runs-on: macos-latest
+    runs-on: flyci-macos-large-latest-m1
   steps:
   - name: 👀 Checkout repo
     uses: actions/checkout@v4
```
# 500 mins/month Free for Public Repos

Since your repo is public, FlyCI offers 500 mins/month of free M1 runner usage with the `flyci-macos-large-latest-m1` runner for public projects.

Don't hеsitate to contact us in case the free tier doesn't suit your needs or you experience any issues with the runners. Our team is here to support you!

Best Regards,
Veselina Radeva
Product Manager at [FlyCI](https://flyci.net/)
