+++ paper/paper.md
@@ -47,7 +47,7 @@ Despite this, quantitative frameworks for modelling the relationships between ev
 
 autoStreamTree employs the Python networkx library [Hagberg2008] to parse geospatial input (i.e., large stream networks) into a graph structure with stream segments as edges, sampling locations as endpoints, and river junctions as nodes. Sample data comprise a tab-delimited table of latitude/longitude coordinates, genome-wide variant data in VCF format, and (optionally) a tab-delimited population map. The data structure ‘graph’ on which downstream computations are performed is built as follows: 1) Sample points are ‘snapped’ to nearest river network nodes (i.e., defining endpoints); 2) Shortest paths are identified between each set of endpoints [@Dijkstra1959]; and 3) A minimal network of original geometries, with contiguous edges derived by joining individual segments with junctions (nodes) retained that fulfill shortest paths.
 
-Pairwise genetic distances from VCF-formatted genotypes are de-rived among individuals, sites, or populations (via a priori user-specifications). Options for sequence- and frequency-based statistics are provided (`-d/--dist`). Mantel tests are available to quantify correla-tions among genetic/ hydrologic distance matrices. The primary workflow is a least-squares procedure analogous to that used to compute branch lengths within a neighbor-joining phylogenetic tree [@Kalinowski2008]. The procedure fits components of the genetic matrix to *k*-segments in a network, such that fitted distance values (*r*) for each segment separating two populations will sum to the observed pairwise matrix value. This provides a distance (*r*~k~) for each of *k*-segments as the genetic distance ‘explained’ by that segment.
+Pairwise genetic distances from VCF-formatted genotypes are derived among individuals, sites, or populations (via a priori user-specifications). Options for sequence- and frequency-based statistics are provided (`-d/--dist`). Mantel tests are available to quantify correlations among genetic/ hydrologic distance matrices. The primary workflow is a least-squares procedure analogous to that used to compute branch lengths within a neighbor-joining phylogenetic tree [@Kalinowski2008]. The procedure fits components of the genetic matrix to *k*-segments in a network, such that fitted distance values (*r*) for each segment separating two populations will sum to the observed pairwise matrix value. This provides a distance (*r*~k~) for each of *k*-segments as the genetic distance ‘explained’ by that segment.
 
 Workflow steps are controlled through the command-line interface (`-r/--run`), with results as plain text tables, and plots via the seaborn package [@Waskom2021]. Fitted distances are added as annotations to an exported geodatabase.
 
@@ -69,7 +69,7 @@ We examined variation in per-locus fitted distances as a function of environment
 
 ## Results and comparison
 
[…]
