### 2024-01-23 0b36ec1
Fixes from issue #6
