### 2024-01-23 8cf0beb
Fixes from issue #6
