+++ paper/paper.bib
@@ -31,6 +31,16 @@
   doi={10.1111/ddi.13160}
 }
 
+@article{Danecek2011,
+  title={The variant call format and VCFtools},
+  author={Danecek, Petr and Auton, Adam and Abecasis, Goncalo and Albers, Cornelis A and Banks, Eric and DePristo, Mark A and Handsaker, Robert E and Lunter, Gerton and Marth, Gabor T and Sherry, Stephen T and McVean, Gilean and Durbin, Richard and 1000 Genomes Analysis Group},
+  journal={Bioinformatics},
+  volume={27},
+  pages={2156--2158},
+  year={2011},
+  doi={10.1093/bioinformatics/btr330},
+}
+
 @article{Dijkstra1959,
   title={A note on two problems in connexion with graphs},
   author={Dijkstra, E W},
