### 2024-01-23 tkchafin/autostreamtree#6 @tkchafin
Thanks for this. I've addressed all that needed correcting and added a link to the data repo (was previously only mentioned in the README).

Re: The USFWS statement -- this is a required boilerplate text due to one of the co-authors being a USFWS employee

Re: State of the field -- this doesn't seem to be a requirement as a sub-heading specifically as far as I can tell, but the current statement of need (especially 2nd paragraph) I think fulfills this requirement
