### 2024-01-23 3ce6cfc
Fix clock_seed issue from #5
