### 2024-01-23 681ca8b
Update ci.yml to add Mac vm
