### 2024-01-23 af931b5
suppress some irrelevant warnings
