### 2024-01-27 tkchafin/autostreamtree#7 @abhishektiwari
[JOSS-REVIEW] autostreamtree install fails due to pysam
Running in Docker container

```
docker run -it --rm condaforge/miniforge3
```

Followed installation instructions,
```
(streamtree) root@79da57b9716d:/# mamba install -c ecoevoinfo -c conda-forge -c bioconda autostreamtree

Looking for: ['autostreamtree']

ecoevoinfo/linux-aarch64 (check zst)                Checked  0.2s
ecoevoinfo/noarch (check zst)                       Checked  0.1s
warning  libmamba Could not parse mod/etag header
warning  libmamba Could not parse mod/etag header
bioconda/linux-aarch64 (check zst)                  Checked  0.1s
bioconda/noarch (check zst)                         Checked  0.2s
ecoevoinfo/linux-aarch64                           125.0 B @   1.3kB/s  0.1s
bioconda/linux-aarch64                             854.0 B @   5.7kB/s  0.2s
ecoevoinfo/noarch                                    1.4kB @   7.1kB/s  0.2s
conda-forge/noarch                                  13.3MB @  27.5MB/s  0.5s
conda-forge/linux-aarch64                           10.2MB @  16.0MB/s  0.7s
bioconda/noarch                                      5.1MB @   4.6MB/s  1.0s

Pinned packages:
  - python 3.10.*

Could not solve for environment specs
The following package could not be installed
└─ autostreamtree is not installable because it requires
   └─ pysam, which does not exist (perhaps a missing channel).
```

**Update**
- Install pysam via pip but still failing to install autostreamtree

```
conda install pip
pip install pysam
```

[JOSS-REVIEW](https://github.com/openjournals/joss-reviews/issues/6160)
