### 2024-01-30 #6160 @abhishektiwari
@tkchafin See my initial feedback

## Installation
At this point I am unable to test the package due to installation issues. I tried installing on Mac, Ubuntu 22.04, Docker images (condaforge/miniforge3:latest ) but no luck. Can you please share guidance on [issue here](https://github.com/tkchafin/autostreamtree/issues/7)?

## Data & Scripts
I see you added reference to data and scripts associated with this work but DOI format seems wrong. Can you update the DOI URL format for archive (https://doi.org/10.17605/OSF.IO/4UXFJ).
