### 2024-01-30 tkchafin/autostreamtree#7 @tkchafin
Hi @abhishektiwari, thanks for documenting the issue. It looks like pysam doesn't have a conda installation for linux-aarch64 (only for linux-64). I'll look into today if I can build an alternate distribution for pysam that will fix the problem.
