### 2024-01-30 tkchafin/autostreamtree#7 @tkchafin
So I looked into it, and there are a lot of compiled dependencies for pysam which also won't support ARM, so fixing that issue just results in more errors. Using an ARM linux you will need to use a workaround as for ARM Macs. In this case you should just be able to add `--platform linux/amd64` to your docker call:
```docker run -it --rm --platform linux/amd64```

This should make it installable for you. I'll go ahead and put a docker image together for this as well, and update the documentation once published
