+++ .github/workflows/conda-update.yml
@@ -1,0 +1,43 @@
+name: Conda Build and Upload
+
+on:
+  release:
+    types: [published]
+
+jobs:
+  build-and-upload:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check Out Code
+        uses: actions/checkout@v2
+      
+      - name: Extract Version from Tag
+        shell: bash
+        run: echo "VERSION=${GITHUB_REF#refs/tags/}" >> $GITHUB_ENV
+
+      - name: Update Version in Conda Recipe
+        shell: bash
+        run: |
+          sed -i "s/\$VERSION/${{ env.VERSION }}/g" conda-recipe/meta.yaml
+
+      - name: Set up Conda
+        uses: conda-incubator/setup-miniconda@v2
+        with:
+          auto-update-conda: true
+          activate-environment: build-env
+          python-version: 3.10
+          miniconda-version: latest
+
+      - name: Build Conda Package
+        shell: bash -l {0}
+        run: |
+          conda install conda-build anaconda-client
+          conda build conda-recipe --output-folder ./build -c ecoevoinfo -c conda-forge -c bioconda
+
+      - name: Upload Conda Package to Anaconda Cloud
+        shell: bash -l {0}
+        env:
+          ANACONDA_API_TOKEN: ${{ secrets.ANACONDA_TOKEN }}
+        run: |
+          anaconda upload --user ecoevoinfo --label main ./build/noarch/*.tar.bz2
+
