+++ .github/workflows/conda-update.yml
@@ -25,7 +25,7 @@ jobs:
         with:
           auto-update-conda: true
           activate-environment: build-env
-          python-version: 3.10
+          python-version: "3.10"
           miniconda-version: latest
 
       - name: Build Conda Package
