### 2024-01-31 0b457f8
merged conda and docker update workflows
