### 2024-01-31 tkchafin/autostreamtree#7 @tkchafin
I added pre-built Docker images you can access like:
```
docker run -it tkchafin/autostreamtree:latest
```

Also added this to the documentation, and the CI/CD workflows will now build and upload to conda and docker on each new release. Let me know if you encounter any further issues, but this is all working on my end so I'll close for now
