### 2024-01-31 tkchafin/autostreamtree#7 @abhishektiwari
Fantastic let me give a try. Should be good now. Feel free to close it for now.
