+++ README.md
@@ -125,7 +125,7 @@ conda install -c ecoevoinfo -c conda-forge -c bioconda autostreamtree
 
 Docker images are also built and hosted for all releases on DockerHub. To run using the latest release:
 ```
-docker run -it tkchafin/autostreamtree:latest
+docker run --platform linux/amd64 -it tkchafin/autostreamtree:latest
 ```
 
 If you require a certain release, you can reference the tag by changing the image name like so:
