### 2024-02-02 tkchafin/autostreamtree#8 @abhishektiwari
[JOSS Review] Publish the package docs as ReadTheDocs
Firstly, I find it really hard to browse the PDF package doc generated by pandoc. Secondly, PDF docs are not automatically generated for every version. So for instance, current version of [autostreamtree](https://github.com/tkchafin/autostreamtree) is `v1.1.3` but [PDF doc](https://github.com/tkchafin/autostreamtree/blob/main/docs/autostreamtree_v1.1.0_2024-01-23.pdf) was generated on back of `v1.1.0`.

I strongly suggest to publish the package docs as [Read The Docs](https://docs.readthedocs.io/en/stable/tutorial/index.html#importing-the-project-to-read-the-docs). Once project created you can use webhooks  to update the docs to latest version.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/6160)
