### 2024-02-02 #6160 @abhishektiwari
@tkchafin Thanks for adding Docker images for the `autoStreamTree`. I was able to quickly validate the functionality and examples.

## Suggested improvement (non-blocking)
1. I think it will be great if you can move the `autoStreamTree` package doc from pandoc generated PDF to [Read the docs](https://github.com/tkchafin/autostreamtree/issues/8). Given docstrings already exists in package files, it will be relatively small change to enable Read the docs and I strongly recommend it. It will improve readability of the package API documentation.

2. `autoStreamTree` is offered as CLI which is great to interact and interface with modular data pipeline tools like Nextflow. This seems right approach in my opinion. That said, given a large number of arguments as input, `autoStreamTree` can benefit by improving the CLI interface. Something like Python [click](https://click.palletsprojects.com/en/8.1.x/) will make this [params](https://github.com/tkchafin/autostreamtree/blob/main/autostreamtree/params.py) implementation more simpler, extendable, and unit testable. In current structure, I am not sure if all permutation/combination of inputs are testable. Again, author has done a great job adding enough automated test to detect regression so I am less fussy about this improvement.
