### 2024-02-02 tkchafin/autostreamtree#8 @tkchafin
Thanks, that's a good idea, I'll look into that and see about tying it in to the existing CI/CD workflow
