### 2024-02-04 tkchafin/autostreamtree#5 @xin-huang
Hello @tkchafin

Thank you for the update.

With the current version in this repository, I can proceed with the example analysis. However, when I used the command listed in the README to install `autostreamtree` from the [ecoevoinfo](https://anaconda.org/ecoevoinfo/autostreamtree) channel:

```
mamba install -c ecoevoinfo -c conda-forge -c bioconda autostreamtree
```

I still got the same error: `ValueError: Seed must be between 0 and 2**32 - 1`.

To resolve this, I had to explicitly specify the version with the following command:

```
mamba install -c ecoevoinfo -c conda-forge -c bioconda autostreamtree==v1.1.3
```

This suggests that the conda channel might need an update to automatically provide the latest version to users.

Besides, I observed that the version on the conda channel is `v1.1.3`, whereas the `setup.py` in this repository lists version `1.0.1`.

https://github.com/tkchafin/autostreamtree/blob/3faf71c0340a8139cef79be4a2a6a7e1827bb074/setup.py#L5

It might be beneficial to synchronize the version in `setup.py` with the conda channel to avoid confusion.
