### 2024-02-04 tkchafin/autostreamtree#6 @xin-huang
Thanks for the update.

To enhance accessibility for readers, I suggest converting the DOI reference `doi: 10.17605/OSF.IO/4UXF` in the manuscript to a clickable hyperlink `[doi: 10.17605/OSF.IO/4UXFJ](https://doi.org/10.17605/OSF.IO/4UXFJ)`. Additionally, it would be beneficial to explicitly mention that the data and scripts are located within the `streamtree` folder for clarity.

I've observed that `original_streamtree_output.xlsx` and `sTree_rda.R` have been added to the project on [OSF](https://osf.io/4uxfj). However, I'm still unsure about the steps required to reproduce the figures presented in your paper.

I presume the `original_streamtree_output.xlsx` file is intended for plotting Figure 1A. Maybe it is the input of `streeToDendrogram`? Could you provide further details or the necessary scripts for recreating this figure?

Similarly, I believe `sTree_rda.R` is used for generating Figures B and C. This script seems to require a specific file: `~/programs/autoStreamTree/out5_perlocLinGst.streamTree.txt`. Sharing this file would greatly facilitate the accurate reproduction of the RDA plots depicted in your figures.
