+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name='autostreamtree',
-    version='1.0.1',
+    version='$VERSION',  # Placeholder
     packages=find_packages(),
     entry_points={
         'console_scripts': [
