### 2024-02-07 7a80889
Update generate_docs.sh
