+++ autostreamtree/sequence.py
@@ -1,271 +1,0 @@
-from typing import List
-
-
-def decode(gt: tuple, ref: str, alts: list, as_iupac: bool = False,
-           as_tuple: bool = False, as_list: bool = False) -> str:
-    """
-    Decode a genotype from a VCF file and return the corresponding DNA sequence
-
-    Args:
-        gt (tuple): The genotype as a tuple of integers representing the
-                    indices of the reference and alternate alleles.
-        ref (str): The reference allele.
-        alts (list): A list of alternate alleles.
-        as_iupac (bool): Whether to return the consensus IUPAC symbol if the
-                         genotype is heterozygous. Default is False.
-        as_tuple (bool): Whether to return the result as a tuple of two strings
-                         Default is False.
-        as_list (bool): Whether to return the result as a list of two strings.
-                        Default is False.
-
-    Returns:
-        ret (str): The decoded genotype as a string. By default, this is a
-                   string in the format "ref/alt", but the format can be
-                   customized using the optional arguments.
-    """
-    # initialize return value
-    ret = [None, None]
-
-    # if genotype is missing, set to "N"
-    if gt[0] is None or gt[1] is None:
-        ret = ["N", "N"]
-    else:
-        # decode each allele
-        if gt[0] == 0:
-            ret[0] = ref
-        else:
-            ret[0] = alts[gt[0] - 1]
-        if gt[1] == 0:
-            ret[1] = ref
-        else:
-            ret[1] = alts[gt[1] - 1]
-
-    # apply optional formatting
-    if as_iupac:
-        return dna_consensus("/".join(ret))
-    elif as_tuple:
-        return tuple(ret)
-    elif as_list:
-        return ret
-    else:
-        return "/".join(ret)
-
-
-def get_nuc_freqs(seqs: dict, ploidy: int) -> list:
-    """
-    Compute the nucleotide frequencies of a set of DNA sequences.
-
-    Args:
-        seqs (dict): A dictionary of DNA sequences. The keys are sample names
-                     and the values are strings of nucleotides.
-        ploidy (int): The ploidy of the sequences. If ploidy=1, ambiguities
-                      will be skipped. If ploidy=2, ambiguities will be
-                      resolved.
-
-    Returns:
-        freqs (list): A list of dictionaries, where each dictionary contains
-                      the nucleotide frequencies for a single position in the
-                      sequences.
-
-    Raises:
-        ValueError: If any sequence is empty or contains only invalid
-                    characters, or if sequences are not all the same length.
-    """
-    if not all(seqs):
-        raise ValueError("One or more sequences are empty.")
-
-    sequence_length = len(seqs[list(seqs.keys())[0]])
-    if not all(len(seq) == sequence_length for seq in seqs.values()):
-        raise ValueError("Sequences are not all the same length.")
-
-    freqs = []
-    for loc in range(sequence_length):
-        allnucs = ""
-        for samp in seqs.keys():
-            allnucs += dna_consensus(seqs[samp][loc]).lower()
-
-        badchars = ["?", "-", "n"]
-        if ploidy == 1:
-            badchars += ["r", "y", "s", "w", "k", "m", "b", "d", "h", "v"]
-        allnucs = ''.join([nuc for nuc in allnucs if nuc not in badchars])
-
-        if ploidy == 2:
-            iupacs = ["r", "y", "s", "w", "k", "m", "b", "d", "h", "v"]
[…]
