+++ generate_docs.sh
@@ -4,10 +4,9 @@
 PACKAGE_NAME="autostreamtree"
 
 # Generate documentation in HTML
-pdoc --html --output-dir ./docs $PACKAGE_NAME
-
-# Assuming pdoc generates a folder with the package name, move it to the root of /docs
-mv ./docs/$PACKAGE_NAME/* ./docs/
+pdoc3 --html --output-dir ./docs ./$PACKAGE_NAME
+mv ./docs/autostreamtree/* ./docs
+rm -rf ./docs/autostreamtree
 
 echo "HTML documentation generated in ./docs/"
 
