+++ conda-recipe/meta.yaml
@@ -34,6 +34,7 @@ requirements:
   - mantel
   - pysam
   - arrow
+  - pdoc3
   - scikit-learn >1
 
 build:
