+++ .github/workflows/conda-docker-update.yml
@@ -18,9 +18,6 @@ jobs:
           echo "VERSION=${VERSION}" >> $GITHUB_ENV
           echo "Stripped version: ${VERSION}"
 
-      - name: Set CI Version Environment Variable
-        run: echo "CI_VERSION=${{ env.VERSION }}" >> $GITHUB_ENV
-
       - name: Update Version in Conda Recipe
         shell: bash
         run: |
@@ -36,6 +33,8 @@ jobs:
 
       - name: Build Conda Package
         shell: bash -l {0}
+        env:
+          CI_VERSION: ${{ env.VERSION }}
         run: |
           conda install conda-build anaconda-client
           conda build conda-recipe --output-folder ./build -c ecoevoinfo -c conda-forge -c bioconda
@@ -69,3 +68,5 @@ jobs:
           tags: |
             tkchafin/autostreamtree:latest
             tkchafin/autostreamtree:${{ env.VERSION }}  # Use stripped version
+          build-args: |
+            CI_VERSION=${{ env.VERSION }}
