### 2024-02-07 tkchafin/autostreamtree#8 @tkchafin
I've integrated the documentation build as HTML into the CI workflow, which will update on each commit. Documentation will be viewable at: https://tkchafin.github.io/autostreamtree/
Also adding a note with the link to the README
