+++ autostreamtree/params.py
@@ -1,0 +1,335 @@
+import sys
+import getopt
+
+
+# Object to parse command-line arguments
+class parseArgs():
+    def __init__(self):
+        # Define options
+        try:
+            options, _ = getopt.getopt(
+                sys.argv[1:], 'hs:i:r:p:d:a:lw:o:gP:L:Scn:G:v:C:O:',
+                ["shp=", "help", "input=", "run=", "pop=", "popmap=", "dist=",
+                 "agg_method=", "het", "genmat=", "snp", "snps", "msat",
+                 "msats", "log", "and_log", "iterative", "weight=", "out=",
+                 "method=", "plots", "plot", "perm=", "phased", "median",
+                 "diploid", "geopop", "geopops", "global_het", "haploid",
+                 "loc_agg=", "pop_agg=", "sdist_agg=", "clusterpop",
+                 "epsilon=", "min_samples=", "sclusterpop", "network=",
+                 "overwrite", "reachid_col=", "length_col=", "coercemat",
+                 "locmatdir=", "vcf=", "concat=", "edge_list=", "gdf_out=",
+                 "seed="]
+            )
+        except getopt.GetoptError as err:
+            print(err)
+            self.display_help(
+                "\nExiting because getopt returned non-zero exit status.")
+        # Default values for params
+        # Input params
+        self.shapefile = None
+        self.locmatdir = None
+        self.geodb = None
+        self.vcf = None
+        self.concat = "none"
+        self.run = "ALL"
+        self.network = None
+        self.pop = False
+        self.overwrite = False
+        self.geopop = False
+        self.clusterpop = False
+        self.sclusterpop = False
+        self.output_driver = "GPKG"
+        self.dist = "LINFST"
+        self.het = False
+        self.genmat = None
+        self.log = False
+        self.and_log = False
+        self.iterative = True
+        self.weight = "CSE67"
+        self.permutations = 1000
+        self.coercemat = False
+        self.method = "PEARSON"
+        self.edge_list = None
+        self.plots = False
+        self.out = "out"
+        self.median = False
+        self.ploidy = 2
+        self.global_het = False
+        self.loc_agg = "ARITH"
+        self.pop_agg = "ARITH"
+        self.sdist_agg = "ARITH"
+        self.seed = None
+
+        self.reachid_col = "HYRIV_ID"
+        self.length_col = "LENGTH_KM"
+
+        # dbscan Options
+        self.min_samples = 1
+        self.epsilon = 20
+
+        # First pass to see if help menu was called
+        for o, a in options:
+            if o in ("-h", "-help", "--help"):
+                self.display_help("Exiting because help menu was called.")
+
+        # Second pass to set all args.
+        for opt, arg in options:
+            arg = arg.strip()
+            opt = opt.replace("-", "")
+
+            if opt in ('s', 'shp'):
+                self.shapefile = arg
+            elif opt in ('h', 'help'):
+                pass
+            elif opt in ('i', 'coords'):
+                self.geodb = arg
+            elif opt in ("C", "concat"):
+                arg_lower = str(arg).lower()
+                if arg_lower in ("none", "all", "loc"):
+                    self.concat = arg_lower
+                else:
+                    self.display_help(f"Invalid option {arg_lower} \
+                                      for option <--concat>")
+            elif opt in ('v', 'vcf'):
+                self.vcf = arg
+            elif opt == "seed":
+                self.seed = int(arg)
+            elif opt in ('r', 'run'):
+                self.run = arg.upper()
+                valid_runs = ["ALL", "GENDIST", "IBD", "STREAMDIST",
+                              "STREAMTREE", "DISTANCES", "RUNLOCI"]
+                if self.run not in valid_runs:
+                    self.display_help(f"Invalid option {self.run} \
[…]
