+++ .github/workflows/conda-docker-update.yml
@@ -39,7 +39,7 @@ jobs:
       - name: Build Conda Package
         shell: bash -l {0}
         run: |
-          conda install conda-build anaconda-client
+          conda install conda-build anaconda-client conda-verify
           conda build conda-recipe --output-folder ./build -c ecoevoinfo -c conda-forge -c bioconda
 
       - name: Upload Conda Package to Anaconda Cloud
