### 2024-02-07 #6160 @tkchafin
@abhishektiwari Thanks for the suggestions, I've moved the documentation to a navigable HTML hosted in GitHub Pages, since this was easy to integrate into the existing CI workflow. This should now be re-built and kept up to date on each commit: https://tkchafin.github.io/autostreamtree/. For now I'll put switching to click as an enhancement as it might be a while before I can get to it, but if it simplifies getting higher test coverage it sounds like a good move
