+++ .github/workflows/conda-docker-update.yml
@@ -32,8 +32,8 @@ jobs:
       - name: Build Conda Package
         shell: bash -l {0}
         run: |
-          conda install conda-build anaconda-client
-          conda build conda-recipe --output-folder ./build -c ecoevoinfo -c conda-forge -c bioconda
+          conda install conda-build anaconda-client conda-verify
+          conda build -c ecoevoinfo -c conda-forge -c biocondaconda-recipe --output-folder ./build 
 
       - name: Upload Conda Package to Anaconda Cloud
         shell: bash -l {0}
