+++ .github/workflows/conda-docker-update.yml
@@ -32,14 +32,18 @@ jobs:
           python-version: "3.10"
           miniconda-version: latest
 
-      - name: Build Conda Package
-        shell: bash
+      - name: Install Dependencies
+        shell: bash -l {0} 
         run: |
-          conda install -y conda-build>=24.1.0 anaconda-client conda-verify
+          conda install conda-build>=24.1.0 anaconda-client conda-verify
+
+      - name: Build Conda Package
+        shell: bash -l {0} 
+        run: |
           conda build conda-recipe --output-folder ./build -c ecoevoinfo -c conda-forge -c bioconda 
 
       - name: Upload Conda Package to Anaconda Cloud
-        shell: bash
+        shell: bash -l {0}
         env:
           ANACONDA_API_TOKEN: ${{ secrets.ANACONDA_TOKEN }}
         run: |
