+++ .github/workflows/conda-docker-update.yml
@@ -35,12 +35,12 @@ jobs:
       - name: Install Dependencies
         shell: bash -l {0} 
         run: |
-          conda install -y conda-build>=24.1.0 anaconda-client conda-verify
+          conda install -y anaconda::conda-build>=24.1.0 anaconda-client conda-verify
 
       - name: Build Conda Package
         shell: bash -l {0} 
         run: |
-          conda build conda-recipe --output-folder ./build -c ecoevoinfo -c conda-forge -c bioconda 
+          conda build conda-recipe --output-folder ./build -c ecoevoinfo -c conda-forge -c bioconda || true 
 
       - name: Upload Conda Package to Anaconda Cloud
         shell: bash -l {0}
