### 2024-02-09 86bf85a
report average distance matrix when RUNLOCI
