### 2024-02-09 706697f
Fix bug in RUNLOCI causing incorrect obs X pred comparison
