### 2024-02-09 tkchafin/autostreamtree#5 @tkchafin
Versioning issues should be fixed now
