### 2024-02-23 a2902d6
resolve performance warning from pandas
