+++ paper/paper.md
@@ -65,7 +65,7 @@ To demonstrate autoStreamTree, we employed existing SNP data for Speckled Dace (
 
 Stream networks were parsed directly as a minimal sub-graph from RiverATLAS, which contains various local-scale environmental/ hydrological features as annotations (i.e., physiography, climate, land-cover, geology, anthropogenic effects) [@Linke2019]. Genetic distances were computed globally and per-locus among sites as linearized *F*~ST~ [@Weir1984] (=*F*~ST~/1-*F*~ST~). To compare with @Kalinowski2008, we used unweighted least-squares, iterative negative distance correction, and replicated analyses using linearized *F*~ST~ independently recalculated in Adegenet [@Jombart2008].
 
-We examined variation in per-locus fitted distances as a function of environmental and anthropogenic covariates, carried over as annotations to RiverATLAS. We reduced N=281 hydro-environmental RiverATLAS attributes using forward-selection in adeSpatial [@Dray2018], after first removing variables showing significant pairwise correlations. Remaining selected variables were employed in partial-redundancy analysis (pRDA; to ‘partial-out’ hydrologic distance), with outliers as stream segments/SNPs with loadings +/-3 standard deviations from the mean.
+We examined variation in per-locus fitted distances as a function of environmental and anthropogenic covariates, carried over as annotations to RiverATLAS. We reduced N=281 hydro-environmental RiverATLAS attributes using forward-selection following the implementation used in adeSpatial [@Dray2018], after first removing variables which were invariant, containing missing values, exhibiting pairwise correlations (\(r\)) over 0.7, or having Variance Inflation Factor (VIF) > 3. Remaining selected variables were used in redundancy analysis (RDA) to visualize variation in fitted distances as a function of environmental factors.
 
 ## Results and comparison
 
