### 2024-02-26 tkchafin/autostreamtree#6 @tkchafin
Sorry this is taking a while, I realized that the `sTree_rda.R` script sourced external functions which were not provided, and I also can't find. To improve reproducibility, I am writing new code for the analyses in Figure 1, which will be available as a nextflow pipeline in a new data repository for the paper. This also provided an opportunity to implement more robust feature selection for the demonstrative RDA, which was somewhat arbitrary in the previous iteration.

New data repo: https://doi.org/10.17605/OSF.IO/9BKGR
