### 2024-03-02 #6160 @tkchafin
Sorry these revisions took a while! In trying to comply with the suggestion in [#6](https://github.com/tkchafin/autostreamtree/issues/6) to make the analysis behind Figure 1 reproducible, I discovered that the R code I had uploaded was calling some custom externally sourced functions that I couldn't find, so I wrote new code for the analysis and plotting. This was also a good chance to make the feature selection more robust (as this was somewhat arbitrary before), as well as to show how to string all of the analysis steps together with autostreamtree as an automated workflow. The new data repository should be self-contained, and has a NextFlow pipeline implementing the analyses used in the case study. Links should all be updated now in the README and paper, so after that is regenerated I think I've addressed all of the issues brought up so far.
