### 2024-03-02 tkchafin/autostreamtree#6 @xin-huang
I tried to run the command:

```
nextflow run run_streamtree.nf
```

However, I got an `AttributeError`:

```
AttributeError: module 'pandas' has no attribute '__version__'
```

Here is the output from `nextflow`:

```
N E X T F L O W  ~  version 23.10.1
Launching `run_streamtree.nf` [sad_ptolemy] DSL2 - revision: 393a6fefa9
[-        ] process > STREAMTREE -
[-        ] process > RUNLOCI    -
[-        ] process > RDA        -
[-        ] process > PLOT       -

[-        ] process > STREAMTREE [  0%] 0 of 1
[-        ] process > RUNLOCI    [  0%] 0 of 1
[-        ] process > RDA        -
[-        ] process > PLOT       -

executor >  local (2)
[24/2256e9] process > STREAMTREE [  0%] 0 of 1
[34/4e06f1] process > RUNLOCI    [  0%] 0 of 1
[-        ] process > RDA        -
[-        ] process > PLOT       -

executor >  local (2)
[24/2256e9] process > STREAMTREE [  0%] 0 of 1
[34/4e06f1] process > RUNLOCI    [  0%] 0 of 1
[-        ] process > RDA        -
[-        ] process > PLOT       -
ERROR ~ Error executing process > 'STREAMTREE'

Caused by:
  Process `STREAMTREE` terminated with an error exit status (1)

Command executed:

  bcftools index spd.vcf.gz
  autostreamtree \
      -s shapefiles/*.shp \
      -v spd.vcf.gz \
      -i global.coords \
      -d "LINFST" \
      --run "ALL" \
      -p global.popmap \
      -O "GPKG" \
      --out "global.ALL" \
      --and_log \
      -w CSE67 \
      --reachid_col HYRIV_ID

Command exit status:
  1

Command output:
  (empty)

Command error:
  Traceback (most recent call last):
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/bin/autostreamtree", line 7, in <module>
      from autostreamtree.cli import main
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/autostreamtree/cli.py", line 4, in <module>
      import geopandas as gpd
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/geopandas/__init__.py", line 1, in <module>
      from geopandas._config import options
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/geopandas/_config.py", line 109, in <module>
      default_value=_default_use_pygeos(),
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/geopandas/_config.py", line 95, in _default_use_pygeos
      import geopandas._compat as compat
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/geopandas/_compat.py", line 17, in <module>
      PANDAS_GE_14 = Version(pd.__version__) >= Version("1.4.0rc0")
  AttributeError: module 'pandas' has no attribute '__version__'

Work dir:
  /test_autostreamtree/work/24/2256e9180bd49e307577af6aa13cea

Tip: view the complete command output by changing to the process work dir and entering the command `cat .command.out`

 -- Check '.nextflow.log' file for details
WARN: Killing running tasks (1)

executor >  local (2)
[24/2256e9] process > STREAMTREE [100%] 1 of 1, failed: 1 ✘
[34/4e06f1] process > RUNLOCI    [200%] 2 of 1, failed: 2 ✘
[-        ] process > RDA        -
[-        ] process > PLOT       -
ERROR ~ Error executing process > 'STREAMTREE'

Caused by:
  Process `STREAMTREE` terminated with an error exit status (1)

Command executed:

  bcftools index spd.vcf.gz
  autostreamtree \
      -s shapefiles/*.shp \
      -v spd.vcf.gz \
      -i global.coords \
      -d "LINFST" \
      --run "ALL" \
      -p global.popmap \
      -O "GPKG" \
      --out "global.ALL" \
      --and_log \
      -w CSE67 \
      --reachid_col HYRIV_ID

Command exit status:
  1

Command output:
  (empty)

Command error:
  Traceback (most recent call last):
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/bin/autostreamtree", line 7, in <module>
      from autostreamtree.cli import main
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/autostreamtree/cli.py", line 4, in <module>
      import geopandas as gpd
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/geopandas/__init__.py", line 1, in <module>
      from geopandas._config import options
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/geopandas/_config.py", line 109, in <module>
      default_value=_default_use_pygeos(),
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/geopandas/_config.py", line 95, in _default_use_pygeos
      import geopandas._compat as compat
    File "/test_autostreamtree/work/conda/autostreamtree-13e80491fdd1b639f2d272be45f02b9d/lib/python3.10/site-packages/geopandas/_compat.py", line 17, in <module>
      PANDAS_GE_14 = Version(pd.__version__) >= Version("1.4.0rc0")
  AttributeError: module 'pandas' has no attribute '__version__'

Work dir:
  /test_autostreamtree/work/24/2256e9180bd49e307577af6aa13cea

Tip: view the complete command output by changing to the process work dir and entering the command `cat .command.out`

 -- Check '.nextflow.log' file for details
```
