### 2024-03-04 tkchafin/autostreamtree#6 @tkchafin
I think this is caused by my not having been specific enough with the versions in my conda recipes. I'll tweak and see if I can fix
