### 2024-03-04 tkchafin/autostreamtree#6 @tkchafin
I updated the versions in the conda recipe, and changed the setup instructions to create a new conda environment for nextflow, to prevent any conflicts with user's base conda environment. Also changed the `nextrflow.config` to by default point to the subset file for the demo, as running the RDA on the full dataset takes a while.
