### 2024-03-05 tkchafin/autostreamtree#6 @xin-huang
I can run the first two processes `STREAMTREE` and `RUNLOCI`

However, for the process RDA, I got a `TypeError`:

```
Traceback (most recent call last):
    File "/test_autostreamtree/workflow/bin/stree_rda.py", line 333, in <module>
      main(args)
    File "/test_autostreamtree/workflow/bin/stree_rda.py", line 21, in main
      gdf = aggregate_by_edge_id(gdf)
            ^^^^^^^^^^^^^^^^^^^^^^^^^
    File "/test_autostreamtree/workflow/bin/stree_rda.py", line 122, in aggregate_by_edge_id
      agg_funcs = {
                  ^
    File "/test_autostreamtree/workflow/bin/stree_rda.py", line 123, in <dictcomp>
      col: 'sum' if col == 'LENGTH_KM' else 'mean' if np.issubdtype(df[col].dtype, np.number) else most_frequent
                                                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    File "/test_autostreamtree/workflow/work/conda/ecopy-24c35b914e535e3adc83ae7e1e23e2df/lib/python3.11/site-packages/numpy/core/numerictypes.py", line 417, in issubdtype
      arg1 = dtype(arg1).type
             ^^^^^^^^^^^
  TypeError: Cannot interpret '<geopandas.array.GeometryDtype object at 0x153816b0a410>' as a data type
```

Besides, I need to make `bin/stree_rda.py` and `bin/composite_plot.py` executable. I think you should also mention this in [README](https://osf.io/rdp83).
