### 2024-03-05 tkchafin/autostreamtree#6 @tkchafin
I think this is a numpy version issue. I adjusted the env/*.yml files
again and also added a note to README on making the scripts executable
