### 2024-03-05 tkchafin/autostreamtree#6 @xin-huang
The problem still exists with `numpy>=1.26`
