### 2024-03-06 tkchafin/autostreamtree#6 @tkchafin
I think I fixed it now... I forgot to update something after doing my testing..
