### 2024-03-06 tkchafin/autostreamtree#6 @xin-huang
Yes, it works now. Thanks.

Because now the default data is `variants/subset_100_snps.vcf.gz`, you can edit the last sentence in [README](https://osf.io/rdp83)
