### 2024-03-22 #6160 @crvernon
:wave: Great job on this @tkchafin !

The following are changes that I need you to address in the paper:

- [ ] LINE 83:  Since we are an international journal, please state the country designation after the "Colorado River"
- [ ] LINE 96:  Only use single parenthesis around "r"
- [ ] Figure 1 caption:  Are you missing a "the" in "...highlighted according to [the] most closely..."?
- [ ] LINE 116:  spacing should be consistent, either include or not, for  "...environmental/ spatial..." before and after the "/"
- [ ] LINE 131:  your hyperlink to the DOI link is missing something and thus is not formatting correctly.
- [ ] LINE 164:  your "*" are not translating and "R" in the genus should be capitalized as well as the "N" and "A" in North America

The paper is a bit long for what we usually accept in JOSS, but I do like the framing so let's see if we can work with it the way it is.  Once you address the above, we will move on to the next stages of getting you published!  Thanks!
