+++ paper/paper.bib
@@ -100,7 +100,7 @@
 }
 
 @misc{Mussmann2018,
-  title={Diversification across a dynamic landscape: Phylogeography and riverscape genetics of speckled dace (*Rhinichthys osculus*) in Western North America},
+  title={Diversification across a dynamic landscape: Phylogeography and riverscape genetics of speckled dace (\*Rhinichthys osculus\*) in Western North America},
   author={Mussmann, Steven Michael},
   publisher={University of Arkansas},
   journal={Ph.D. Dissertation},
