+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{CampbellGrant2007,
-  title={Living in the branches: population dynamics and ecological processes in dendritic networks},
+  title={Living in the branches: Population dynamics and ecological processes in dendritic networks},
   author={Campbell Grant, E H and Lowe, W H and Fagan, W F},
   journal={Ecology Letters},
   volume={10},
@@ -100,7 +100,7 @@
 }
 
 @misc{Mussmann2018,
-  title={Diversification across a dynamic landscape: Phylogeography and riverscape genetics of speckled dace (Rhinichthys osculus) in Western North America},
+  title={Diversification across a dynamic landscape: Phylogeography and riverscape genetics of speckled dace \(*Rhinichthys osculus*\) in Western North America},
   author={Mussmann, Steven Michael},
   publisher={University of Arkansas},
   journal={Ph.D. Dissertation},
