### 2024-03-23 #6160 @crvernon
Thanks @tkchafin , you still need to remove the extra space in "...environmental/ spatial..." to simply make it read "...environmental/spatial..." per the following comment:

>  LINE 116: spacing should be consistent, either include or not, for "...environmental/ spatial..." before and after the "/"
