### 2024-03-23 #6160 @crvernon
Alright @tkchafin , looking good!  - we are almost there! Next is just setting up the archive for your new release.

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

- [ ]  Please respond with the DOI of the archived version here
